2022 Third class requirement 1, Add elective class in school year setup 2, Can re-register elective class 3, Can withdraw elective class only

+++ chineseschool/app/models/withdraw_request.rb
@@ -88,18 +88,37 @@ class WithdrawRequest < ActiveRecord::Base
       student_status_flag = withdraw_request_detail.student.student_status_flag_for current_school_year
       unless student_status_flag.nil?
         withdrawal_record.registration_date = student_status_flag.last_status_change_date
-        student_status_flag.registered = false
+        # If only withdraw elective class, then not update registered = false
+        unless withdraw_request_detail.elective_class_only == 'Y'
+          student_status_flag.registered = false
+        end
         student_status_flag.last_status_change_date = self.updated_at
         student_status_flag.save!
       end
 
       # Move class assignment data to withdrawal record and destroy the class assignment
       student_class_assignment = withdraw_request_detail.student.student_class_assignment_for current_school_year
-      unless student_class_assignment.nil?
-        withdrawal_record.grade_id = student_class_assignment.grade_id
-        withdrawal_record.school_class_id = student_class_assignment.school_class_id
+      # If only withdraw elective class, then not destroy student_class_assignment, just update student_class_assignment.elective_class_id = 0
+      if withdraw_request_detail.elective_class_only == 'Y'
+        withdrawal_record.grade_id = 0
+        withdrawal_record.school_class_id = 0
         withdrawal_record.elective_class_id = student_class_assignment.elective_class_id
-        student_class_assignment.destroy
+        student_class_assignment.elective_class_id = 0
+        student_class_assignment.save!
+
+        # Set elective_class_id = 0 for registration_preference
+        registration_preference = withdraw_request_detail.student.registration_preference_for(current_school_year)
+        unless registration_preference.nil?
+          registration_preference.elective_class_id = 0
+          registration_preference.save!
+        end
+      else
+        unless student_class_assignment.nil?
+          withdrawal_record.grade_id = student_class_assignment.grade_id
+          withdrawal_record.school_class_id = student_class_assignment.school_class_id
+          withdrawal_record.elective_class_id = student_class_assignment.elective_class_id
+          student_class_assignment.destroy
+        end
       end
       withdrawal_record.save!
